抽离layout初始化模块
